requirements.txt
streamlit
pandas
folium
streamlit-folium
geopandas
openpyxl
pyyaml
bcrypt
streamlit-authenticator
